Battleforce Angel, Magic Initiate, Keen Mind
For some reason the writers of Guildmaster’s Guide to Ravnica forgot that giving 4d8 Radiant Damage to every melee attack from a given celestial creature (in this case, the Battleforce Angel) means that that creature has the Angelic Weapons trait. This is important as the Angelic Weapons trait is still applicable when you use Shapechange.
Magic Initiate needed to be revised in order to be sensible with the understanding that only creatures other than those stated in the Reincarnate spell have innate spellcasting abilities (a certain number of casts of a particular spell per day, non-transferable), due to the fact that every class and School of Magic uses transferable spellcasting (either spell slots or sorcery points). We also can’t use these weird rules of changing what your spellcasting ability is.

Keen Mind would have undermined the concept of “maximum prepared spells”, but we kept the language effectively the same and the inherent meaning. This of course requires a clarification that, in order for the feat to mean the same thing, the main limitation on Sorcerers (outside of the length of their spell list) of the number of KNOWN spells, would need to mostly go away in the face of being able to learn new spells in the sorcerer spell list on the fly. This would not allow them to beat the maximum prepared spell limit for other classes they are granted levels in (Druid and Cleric). We don’t have to consider how this effects Sorcerers learning other spells through Schools of Magic or spell scrolls because the magic of a sorcerer innately makes it impossible for them to do so.